09_agents_as_tool/pyproject.toml
[project]
name = "09-agents-as-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "grpcio<1.72.1",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.0",
]
